Convert some rules to be opt-in because they are noisy for non-strict projects.
The rules that are now opt-in are:
- AllPublicDeclarationsHaveDocumentation
- NeverForceUnwrap
- NeverUseForceTry
- NeverUseImplicitlyUnwrappedOptionals

The rules that identify unsafe patterns (force unwrap, force try, implicitly unwrapped optionals) aren't able to understand the context of the usage, which means the use might be safe and the rule can't tell.

Projects that wish to be more strict can turn on these rules in their configuration, to get back to the original behavior. There is also no change for projects that are already using a configuration file.