Handle #if clauses inside of switch statements.
The special IfConfigDecl, aka `#if`, can exist in the `cases` of a SwitchStmtSyntax because the `#if` "wraps" one or more switch cases. These weren't being considered for the purposes of indenting the cases. There were also no breaks between the switch statement's left brace and `#if` when `#if` was the first "case" which caused malformed code.

I have opted to treat `indentSwitchCaseLabels` as also indenting the `#if` cases in a switch statement, because that makes the `#if` here consistent with other locations.