Apply private and fileprivate rules to UseSynthesizedInitializer.
The linter rule didn't handle the cases when the stored properties of a struct are private or fileprivate. A synthesized memberwise initializer has access level according to the stored properties. If a type had a private stored property and a memberwise initializer without an access level modifier, then the rule was suggesting to remove the initializer even though the synthesized initializer would have private visibility.